+++ packages/bluesky/lib/src/services/entities/content_label_preference.dart
@@ -24,6 +24,7 @@ class ContentLabelPreference with _$ContentLabelPreference {
   const factory ContentLabelPreference({
     @typeKey @Default(appBskyActorDefsContentLabelPref) String type,
     required String label,
+    String? labelerDid,
     required ContentLabelVisibility visibility,
   }) = _ContentLabelPreference;